base refactoring
