+++ config/urls.py
@@ -22,7 +22,7 @@ urlpatterns = [
     # User management
     url(r'^users/', include('infty.users.urls', namespace='users')),
     url(r'^accounts/', include('allauth.urls')),
-    url(r'^api/', include('infty.api.urls')),
+    url(r'^api/', include('infty.api.urls', namespace='api')),
     url(r'^rest-auth/', include('rest_auth.urls')),
     url(r'^rest-auth/registration/', include('rest_auth.registration.urls')),
     url(r'^captcha/', include('captcha.urls')),
