+++ infty/api/v1/core/serializers.py
@@ -13,22 +13,14 @@ from infty.core.models import (
     Comment,
 )
 from infty.transactions.models import (
-    Currency,
-    Transaction,
-    Interaction,
-    TopicSnapshot,
-    CommentSnapshot,
-    HourPriceSnapshot,
-    CurrencyPriceSnapshot,
-    ContributionCertificate
-)
-from infty.users.models import (
-    User,
-    LanguageName
-)
+    Currency, Transaction, Interaction, TopicSnapshot, CommentSnapshot,
+    HourPriceSnapshot, CurrencyPriceSnapshot, ContributionCertificate)
+from infty.users.models import (User, LanguageName)
+
 
 class LangSplitField(serializers.CharField):
     """Langsplit CharField"""
+
     def to_internal_value(self, data):
         return super(LangSplitField, self).to_internal_value(data)
 
@@ -37,7 +29,8 @@ class LangSplitField(serializers.CharField):
 
         if lang and value:
             split = splitter.split(value, title=True)
-            return split.get(lang) or 'languages: {}'.format(list(split.keys()))
+            return split.get(lang) or 'languages: {}'.format(
+                list(split.keys()))
 
         return value
 
@@ -52,7 +45,6 @@ class TypeSerializer(serializers.HyperlinkedModelSerializer):
 
 
 class InstanceSerializer(serializers.HyperlinkedModelSerializer):
-
     class Meta:
         model = Instance
         fields = ('url', 'role', 'description', 'languages')
@@ -62,10 +54,7 @@ class CategoriesField(serializers.RelatedField):
     def to_representation(self, value):
         lang = self.context['request'].query_params.get('lang')
 
-        item = {
-            "id": value.pk,
-            "name": value.name
-        }
+        item = {"id": value.pk, "name": value.name}
 
         if lang:
 
@@ -77,7 +66,6 @@ class CategoriesField(serializers.RelatedField):
 
 
 class UserRepresentation(serializers.CharField):
-
     def to_representation(self, value):
         return {"id": value.pk, "username": value.username}
 
@@ -87,23 +75,26 @@ class TopicSerializer(serializers.HyperlinkedModelSerializer):
     body = LangSplitField(required=False)
     type = serializers.ChoiceField(choices=Topic.TOPIC_TYPES, required=False)
     owner = UserRepresentation(read_only=True)
-    editors = serializers.ReadOnlyField(source='editors.username', read_only=True)
+    editors = serializers.ReadOnlyField(
+        source='editors.username', read_only=True)
     parents = serializers.HyperlinkedRelatedField(
-        many = True,
+        many=True,
         view_name='topic-detail',
         queryset=Topic.objects.all(),
-        required=False
-    )
+        required=False)
 
     class Meta:
         model = Topic
-        fields = ('id', 'url', 'type', 'title', 'body', 'owner', 'editors', 'parents', 'categories', 'languages', 'is_draft', 'blockchain')
+        fields = ('id', 'url', 'type', 'title', 'body', 'owner', 'editors',
+                  'parents', 'categories', 'languages', 'is_draft',
+                  'blockchain')
 
 
 class CommentSerializer(serializers.HyperlinkedModelSerializer):
 
     text = LangSplitField(required=True)
-    topic = serializers.HyperlinkedRelatedField(view_name='topic-detail', queryset=Topic.objects.all())
+    topic = serializers.HyperlinkedRelatedField(
+        view_name='topic-detail', queryset=Topic.objects.all())
     owner = UserRepresentation(read_only=True)
 
     def get_text(self, obj):
@@ -111,42 +102,47 @@ class CommentSerializer(serializers.HyperlinkedModelSerializer):
 
         if lang:
             split = splitter.split(obj.text)
-            return split.get(lang) or 'languages: {}'.format(list(split.keys()))
+            return split.get(lang) or 'languages: {}'.format(
+                list(split.keys()))
 
         return obj.text
 
     class Meta:
         model = Comment
-        fields = ('id', 'url', 'topic', 'text', 'claimed_hours', 'assumed_hours', 'owner', 'languages', 'matched', 'donated', 'remains', 'parent', 'blockchain')
+        fields = ('id', 'url', 'topic', 'text', 'claimed_hours',
+                  'assumed_hours', 'owner', 'languages', 'matched', 'donated',
+                  'remains', 'parent', 'blockchain')
 
 
 class TopicSnapshotSerializer(serializers.HyperlinkedModelSerializer):
 
-    topic = serializers.HyperlinkedRelatedField(view_name='comment-detail', queryset=Comment.objects.all())
+    topic = serializers.HyperlinkedRelatedField(
+        view_name='comment-detail', queryset=Comment.objects.all())
 
     class Meta:
         model = TopicSnapshot
-        fields = ('id', 'created_date', 'topic', 'data', 'blockchain', 'blockchain_tx')
+        fields = ('id', 'created_date', 'topic', 'data', 'blockchain',
+                  'blockchain_tx')
 
 
 class CommentSnapshotSerializer(serializers.HyperlinkedModelSerializer):
 
-    comment = serializers.HyperlinkedRelatedField(view_name='comment-detail', queryset=Comment.objects.all())
+    comment = serializers.HyperlinkedRelatedField(
+        view_name='comment-detail', queryset=Comment.objects.all())
 
     class Meta:
         model = CommentSnapshot
-        fields = ('id', 'created_date', 'comment', 'data', 'blockchain', 'blockchain_tx')
+        fields = ('id', 'created_date', 'comment', 'data', 'blockchain',
+                  'blockchain_tx')
 
 
 class HourPriceSnapshotSerializer(serializers.HyperlinkedModelSerializer):
-
     class Meta:
         model = HourPriceSnapshot
         fields = ('id', 'name', 'base', 'endpoint', 'data')
 
 
 class CurrencyPriceSnapshotSerializer(serializers.HyperlinkedModelSerializer):
-
     class Meta:
         model = CurrencyPriceSnapshot
         fields = ('id', 'name', 'base', 'endpoint', 'data')
@@ -154,8 +150,12 @@ class CurrencyPriceSnapshotSerializer(serializers.HyperlinkedModelSerializer):
 
 class CurrencyListSerializer(serializers.HyperlinkedModelSerializer):
 
-    hour_price = serializers.HyperlinkedRelatedField(view_name='hourpricesnapshot-detail', queryset=HourPriceSnapshot.objects.all())
-    currency_price = serializers.HyperlinkedRelatedField(view_name='currencypricesnapshot-detail', queryset=CurrencyPriceSnapshot.objects.all())
+    hour_price = serializers.HyperlinkedRelatedField(
+        view_name='hourpricesnapshot-detail',
+        queryset=HourPriceSnapshot.objects.all())
+    currency_price = serializers.HyperlinkedRelatedField(
+        view_name='currencypricesnapshot-detail',
+        queryset=CurrencyPriceSnapshot.objects.all())
 
     class Meta:
         model = Currency
@@ -164,8 +164,10 @@ class CurrencyListSerializer(serializers.HyperlinkedModelSerializer):
 
 class InteractionSerializer(serializers.HyperlinkedModelSerializer):
 
-    comment = serializers.HyperlinkedRelatedField(view_name='interaction-detail', queryset=Comment.objects.all())
-    snapshot = serializers.PrimaryKeyRelatedField (queryset=CommentSnapshot.objects.all())
+    comment = serializers.HyperlinkedRelatedField(
+        view_name='interaction-detail', queryset=Comment.objects.all())
+    snapshot = serializers.PrimaryKeyRelatedField(
+        queryset=CommentSnapshot.objects.all())
 
     class Meta:
         model = Interaction
@@ -174,13 +176,17 @@ class InteractionSerializer(serializers.HyperlinkedModelSerializer):
 
 class TransactionCreateSerializer(serializers.HyperlinkedModelSerializer):
 
-    comment = serializers.HyperlinkedRelatedField(view_name='comment-detail', queryset=Comment.objects.all())
-    payment_currency = serializers.PrimaryKeyRelatedField(queryset=Currency.objects.all())
-    payment_sender = serializers.PrimaryKeyRelatedField(queryset=User.objects.all())
+    comment = serializers.HyperlinkedRelatedField(
+        view_name='comment-detail', queryset=Comment.objects.all())
+    payment_currency = serializers.PrimaryKeyRelatedField(
+        queryset=Currency.objects.all())
+    payment_sender = serializers.PrimaryKeyRelatedField(
+        queryset=User.objects.all())
 
     class Meta:
         model = Transaction
-        fields = ('comment', 'payment_amount', 'payment_currency', 'payment_sender')
+        fields = ('comment', 'payment_amount', 'payment_currency',
+                  'payment_sender')
 
     def create(self, validated_data):
         comment = validated_data.get('comment')
@@ -194,7 +200,6 @@ class TransactionCreateSerializer(serializers.HyperlinkedModelSerializer):
             investor=sender,
         )
 
-
         if not tx:
             raise ValidationError('Bad data')
 
@@ -203,37 +208,49 @@ class TransactionCreateSerializer(serializers.HyperlinkedModelSerializer):
 
 class TransactionListSerializer(serializers.HyperlinkedModelSerializer):
 
-    comment = serializers.HyperlinkedRelatedField(view_name='comment-detail', queryset=Comment.objects.all())
-    snapshot = serializers.PrimaryKeyRelatedField (queryset=CommentSnapshot.objects.all())
-    hour_price = serializers.PrimaryKeyRelatedField(queryset=HourPriceSnapshot.objects.all())
-    currency_price = serializers.PrimaryKeyRelatedField(queryset=CurrencyPriceSnapshot.objects.all())
-    payment_currency = serializers.PrimaryKeyRelatedField(queryset=Currency.objects.all())
+    comment = serializers.HyperlinkedRelatedField(
+        view_name='comment-detail', queryset=Comment.objects.all())
+    snapshot = serializers.PrimaryKeyRelatedField(
+        queryset=CommentSnapshot.objects.all())
+    hour_price = serializers.PrimaryKeyRelatedField(
+        queryset=HourPriceSnapshot.objects.all())
+    currency_price = serializers.PrimaryKeyRelatedField(
+        queryset=CurrencyPriceSnapshot.objects.all())
+    payment_currency = serializers.PrimaryKeyRelatedField(
+        queryset=Currency.objects.all())
     payment_recipient = UserRepresentation(read_only=True)
     payment_sender = UserRepresentation(read_only=True)
 
     class Meta:
         model = Transaction
         fields = ('url', 'comment', 'snapshot', 'hour_price', 'currency_price',
-            'payment_amount', 'payment_currency', 'payment_recipient',
-            'payment_sender', 'hour_unit_cost', 'donated_hours', 'matched_hours')
+                  'payment_amount', 'payment_currency', 'payment_recipient',
+                  'payment_sender', 'hour_unit_cost', 'donated_hours',
+                  'matched_hours')
 
 
 class ContributionSerializer(serializers.HyperlinkedModelSerializer):
 
-    transaction = serializers.HyperlinkedRelatedField(view_name='transaction-detail', queryset=Transaction.objects.all())
-    interaction = serializers.HyperlinkedRelatedField(view_name='interaction-detail', queryset=Interaction.objects.all())
-    comment_snapshot = serializers.HyperlinkedRelatedField(view_name='commentsnapshot-detail', queryset=CommentSnapshot.objects.all())
+    transaction = serializers.HyperlinkedRelatedField(
+        view_name='transaction-detail', queryset=Transaction.objects.all())
+    interaction = serializers.HyperlinkedRelatedField(
+        view_name='interaction-detail', queryset=Interaction.objects.all())
+    comment_snapshot = serializers.HyperlinkedRelatedField(
+        view_name='commentsnapshot-detail',
+        queryset=CommentSnapshot.objects.all())
     received_by = UserRepresentation(read_only=True)
 
     class Meta:
         model = ContributionCertificate
-        fields = ('type', 'url', 'interaction', 'transaction', 'received_by', 'comment_snapshot', 'broken', 'parent')
+        fields = ('type', 'url', 'interaction', 'transaction', 'received_by',
+                  'comment_snapshot', 'broken', 'parent')
 
 
 class UserBalanceSerializer(serializers.HyperlinkedModelSerializer):
 
     balance = serializers.SerializerMethodField('matched')
-    contributions = serializers.SerializerMethodField('contribution_certificates')
+    contributions = serializers.SerializerMethodField(
+        'contribution_certificates')
 
     def matched(self, obj):
         return ContributionCertificate.user_matched(obj)
@@ -244,9 +261,8 @@ class UserBalanceSerializer(serializers.HyperlinkedModelSerializer):
         domain = request.META.get('HTTP_HOST') or ''
         endpoint = reverse('contributioncertificate-list')
 
-        return "{}{}{}?received_by={}".format(
-            protocol, domain, endpoint, obj.id
-        )
+        return "{}{}{}?received_by={}".format(protocol, domain, endpoint,
+                                              obj.id)
 
     class Meta:
         model = User
@@ -254,7 +270,6 @@ class UserBalanceSerializer(serializers.HyperlinkedModelSerializer):
 
 
 class LanguageNameSerializer(serializers.HyperlinkedModelSerializer):
-
     class Meta:
         model = LanguageName
         fields = ('lang', 'name', 'enabled')
